Add allowSnakeCase option to dot-notation rule
By default, allowSnakeCase is false.  If set to true, the dot-notation rule will not be enforced if the property has a snake_cased name.  This allows the dot-notation and camelcase rule to be used together in situations where data needs to be prepared to be sent to an external API and property names are outside of the developer's control.